Since only one controller can be loaded per FXML file I've compiled all the "listeners" into a single class. There is a way to split FXML files into several, which would allow different controller classes, but needs to be investigated.
As a wee experiment, the controller handles clicks in the map image and translate them into station selection. So far I'm hard coding the pixel boundaries of each station (by clicking on the map and seeing the console output), but might be a simpler way to do this. Anyway shouldn't take more than an hour to add them all.